Replace lots of $form['blah'] = array('#type' => 'value', '#value' => $blah); elements with $form['#blah'] = $blah; as the former is now deprecated in Drupal 6.
I hope I didn't miss any form_alter or $form_state['values'] conversion,
anyways I tried to be as thorough as possible.

Did not touch the form_altered node form, as it's supposed to have changed
in regards to form element handling, $form['#node'] and all. This will need
to be tested and probably fixed sooner or later.